Globalnet Healthcheck support
Currently, Globalnet only supports Pod to remoteService connectivity.
Submariner healthcheck uses ping/icmp to verify the tunnel status to
remoteClusters. In this PR, Globalnet is enhanced to support ping to
the globalIP allocated to the GatewayNodes, so that health-check can
be performed even for Globalnet deployments.

In a subsequent PR, endpoint info will be updated to use globalIPs as
health-check IPs.

Depends On: https://github.com/submariner-io/submariner-operator/pull/1015
Related to: https://github.com/submariner-io/submariner/issues/970

(cherry picked from commit e523f9a0531c89a43558f89ecdc4b90b495d0438)
