+++ pkg/globalnet/controllers/ipam/globalnet.go
@@ -73,10 +73,19 @@ func (i *Controller) syncServiceRules(service *k8sv1.Service, globalIP string, a
 	return nil
 }
 
-func (i *Controller) syncNodeRules(cniIfaceIP, globalIP string, addRules bool) error {
+func (i *Controller) syncNodeRules(nodeName, cniIfaceIP, globalIP string, addRules bool) error {
 	err := i.updateEgressRulesForResource("Node", cniIfaceIP, globalIP, addRules)
 	if err != nil {
 		return fmt.Errorf("error updating egress rules for Node %s: %v", cniIfaceIP, err)
+	}
+
+	// On the active Gateway Node where this code gets executed, we program icmp ingress rules
+	// to support health-check use-case.
+	if i.gwNodeName == nodeName {
+		err = i.updateIngressRulesForHealthCheck("Node", cniIfaceIP, globalIP, addRules)
+		if err != nil {
+			return fmt.Errorf("error updating healthcheck ingress rules for Node %q: %v", nodeName, err)
+		}
 	}
 
 	return nil
